requirements.txt
torch>=2.0.0
torchvision>=0.15.0
trimesh>=3.20.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
ipywidgets>=8.0
# MaterialAnything dependencies
scikit-learn
opencv-python
matplotlib
imageio
diffusers==0.28.2
einops
transformers
open-clip-torch
omegaconf
accelerate
# Note: pytorch3d and kaolin are complex dependencies that may require special installation
# They are needed for full MaterialAnything functionality but are optional
# Note: shap-e must be installed from GitHub (see setup.sh)

# NeRF dependencies (optional - TensorFlow 1.15 may not work on Apple Silicon)
# tensorflow==1.15  # Very old, may not work on modern systems
# configargparse
# Note: NeRF requires TensorFlow 1.15 and CUDA, which may not be compatible with Apple Silicon
# Consider using modern NeRF implementations like nerfstudio for M-Series chips

